requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
typing-extensions
numpy
requests
python-multipart
starlette
groq
tavily-python
pydantic-settings